second part of the app
second part of the app has been made. it moves one file from one directory to another. first we get list of files in directory. then we are checking if tfile we wish to copy has the same name as file in directory if it has it do not move file or delete it from source if names are different then directory from source is movimg to destination directory. Also, creates file with the same name and extension in the new directory since it is nessecary to have it with the same name and extension.