merge first cmd app and half of move file class
user first must choose source directory this class have if condition if there is any of the file in source directory if there is none finish app with message  that choses directory must not be empty. also check if destination is equal to source directory. if conditions are satysfied, then list are printed and eventualy file move can be made. added static variables which will represent messages.