disabled futures
when open last directories are activated choose source and destination directory will be disabled. Same with the reverse, if  destination or source are choosen, open last directory will be disabled. Note: create file will be useful when json for last directories will be chosen. File can have any name, and needs to implement to have any name. Make a note.